initial commit
Web scraping is a method for extracting data from websites. This notebook demonstrates how to scrape the 100 most active stocks from the Yahoo Finance website. Before scraping a website, it is important to check its terms of service to ensure that web scraping is allowed. Before we get started, we need to install the following Python libraries:
requests- for making HTTP requests to the website BeautifulSoup- for parsing HTML and extracting the needed data pandas- for data manipulation matplotlib- for data visualization. We can install these using pip: pip install requests beautifulsoup4, pandas and matplotlib.